refactor: move rack env building to Socket and make it lazy
