+++ lib/anycable/rpc/handler.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "anycable/rpc/helpers"
-
 require "anycable/rpc/handlers/connect"
 require "anycable/rpc/handlers/disconnect"
 require "anycable/rpc/handlers/command"
@@ -10,7 +8,6 @@ module AnyCable
   module RPC
     # Generic RPC handler
     class Handler
-      include Helpers
       include Handlers::Connect
       include Handlers::Disconnect
       include Handlers::Command
@@ -29,6 +26,25 @@ module AnyCable
       private
 
       attr_reader :commands, :middleware
+
+      def build_socket(env:)
+        AnyCable::Socket.new(env: env)
+      end
+
+      def build_env_response(socket)
+        AnyCable::EnvResponse.new(
+          cstate: socket.cstate.changed_fields,
+          istate: socket.istate.changed_fields
+        )
+      end
+
+      def logger
+        AnyCable.logger
+      end
+
+      def factory
+        AnyCable.connection_factory
+      end
     end
   end
 end
