+++ lib/anycable/rpc/handlers/connect.rb
@@ -7,7 +7,7 @@ module AnyCable
         def connect(request)
           logger.debug("RPC Connect: #{request.inspect}")
 
-          socket = build_socket(env: rack_env(request.env))
+          socket = build_socket(env: request.env)
 
           connection = factory.call(socket)
